Factor out layout creation
